+++ banzai/calibrations.py
@@ -1,16 +1,13 @@
 import logging
 import abc
 import os
-from datetime import datetime, timedelta
 
 import numpy as np
 from astropy.io import fits
 
-from banzai.celery import schedule_stack
-
 from banzai.stages import Stage, MultiFrameStage
 from banzai import dbs, logs, settings
-from banzai.utils import image_utils, stats, fits_utils, qc, date_utils, import_utils, lake_utils
+from banzai.utils import image_utils, stats, fits_utils, qc, date_utils, import_utils
 
 FRAME_CLASS = import_utils.import_attribute(settings.FRAME_CLASS)
 
@@ -253,34 +250,3 @@ def process_master_maker(runtime_context, instrument, frame_type, min_date, max_
     except Exception:
         logger.error(logs.format_exception())
     logger.info("Finished")
-
-
-def schedule_stacking_checks(runtime_context):
-    logger.info('scheduling stacking checks')
-    calibration_blocks = lake_utils.get_calibration_blocks_for_time_range(runtime_context.site,
-                                                                          runtime_context.max_date,
-                                                                          runtime_context.min_date)
-    instruments = dbs.get_instruments_at_site(site=runtime_context.site, db_address=runtime_context.db_address)
-    for instrument in instruments:
-        worker_runtime_context = dict(runtime_context._asdict())
-        worker_runtime_context['enclosure'] = instrument.enclosure
-        worker_runtime_context['telescope'] = instrument.telescope
-        worker_runtime_context['camera'] = instrument.camera
-        blocks_for_calibration = lake_utils.filter_calibration_blocks_for_type(instrument,
-                                                                               worker_runtime_context['frame_type'],
-                                                                               calibration_blocks)
-        if len(blocks_for_calibration) > 0:
-            block_end = datetime.strptime(blocks_for_calibration[0]['end'], date_utils.TIMESTAMP_FORMAT)
-            stack_delay = timedelta(
-                seconds=settings.CALIBRATION_STACK_DELAYS[worker_runtime_context['frame_type'].upper()]
-            )
-            now = datetime.utcnow().replace(microsecond=0)
-            logger.info('before schedule stack for block type {0}'.format(worker_runtime_context['frame_type']))
-            message_delay = block_end - now + stack_delay
-            if message_delay.days < 0:
-                message_delay_in_seconds = 0  # Remove delay if block end is in the past
-            else:
-                message_delay_in_seconds = message_delay.seconds
-            logger.info('scheduling with message delay {0}'.format(str(message_delay_in_seconds)))
-            schedule_stack.apply_async(args=(worker_runtime_context, blocks_for_calibration),
-                                       countdown=message_delay_in_seconds)
