+++ banzai/celery.py
@@ -1,11 +1,13 @@
 import os
 import logging
+from datetime import datetime, timedelta
 
 from celery import Celery
 
 from celery.schedules import crontab
+
 from banzai import settings, dbs, calibrations, logs
-from banzai.utils import date_utils, realtime_utils
+from banzai.utils import date_utils, realtime_utils, lake_utils
 from banzai.context import Context
 from banzai.utils.stage_utils import run
 
@@ -30,7 +32,7 @@ def schedule_calibration_stacking(runtime_context=None, raw_path=None):
     for frame_type in settings.CALIBRATION_IMAGE_TYPES:
         runtime_context_json['frame_type'] = frame_type
         runtime_context = Context(runtime_context_json)
-        calibrations.schedule_stacking_checks(runtime_context)
+        schedule_stacking_checks(runtime_context)
 
 
 @app.task(name='celery.schedule_stack', bind=True, default_retry_delay=RETRY_DELAY)
@@ -91,3 +93,34 @@ def process_image(path, runtime_context_dict):
     except Exception:
         logger.error("Exception processing frame: {error}".format(error=logs.format_exception()),
                      extra_tags={'filename': os.path.basename(path)})
+
+
+def schedule_stacking_checks(runtime_context):
+    logger.info('scheduling stacking checks')
+    calibration_blocks = lake_utils.get_calibration_blocks_for_time_range(runtime_context.site,
+                                                                          runtime_context.max_date,
+                                                                          runtime_context.min_date)
+    instruments = dbs.get_instruments_at_site(site=runtime_context.site, db_address=runtime_context.db_address)
+    for instrument in instruments:
+        worker_runtime_context = dict(runtime_context._asdict())
+        worker_runtime_context['enclosure'] = instrument.enclosure
+        worker_runtime_context['telescope'] = instrument.telescope
+        worker_runtime_context['camera'] = instrument.camera
+        blocks_for_calibration = lake_utils.filter_calibration_blocks_for_type(instrument,
+                                                                               worker_runtime_context['frame_type'],
+                                                                               calibration_blocks)
+        if len(blocks_for_calibration) > 0:
+            block_end = datetime.strptime(blocks_for_calibration[0]['end'], date_utils.TIMESTAMP_FORMAT)
+            stack_delay = timedelta(
+                seconds=settings.CALIBRATION_STACK_DELAYS[worker_runtime_context['frame_type'].upper()]
+            )
+            now = datetime.utcnow().replace(microsecond=0)
+            logger.info('before schedule stack for block type {0}'.format(worker_runtime_context['frame_type']))
+            message_delay = block_end - now + stack_delay
+            if message_delay.days < 0:
+                message_delay_in_seconds = 0  # Remove delay if block end is in the past
+            else:
+                message_delay_in_seconds = message_delay.seconds
+            logger.info('scheduling with message delay {0}'.format(str(message_delay_in_seconds)))
+            schedule_stack.apply_async(args=(worker_runtime_context, blocks_for_calibration),
+                                       countdown=message_delay_in_seconds)
